backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-environ==0.11.2
django-cors-headers==4.6.0
requests==2.32.3
pytest==8.3.5
pytest-django==4.11.1
